+++ CHANGELOG.md
@@ -3,5 +3,8 @@
 ### Removed
 - Removed optional `clippy` dependency
 
+### Added
+- Added `report` method to `Latency`
+
 ## [0.1.0] - 2017-03-24
 - Initial release